style(xls): clear clippy warnings in list-object, chart, comments, and friends
- parenthesize the FEAT11 flags word; collapse nested ifs
- box the legacy CFEx extension (shrinks ParsedExtension)
- drop unneeded Ok(?), same-type casts, and loop/let-else in favor of
  while-let
- derive Default for XlsDataValidationTableOptions; targeted
  type_complexity allows

encryption.rs/data_table.rs lint fixes are left in the tree for the
concurrent xls feature agent that is already editing those files;
litchi-ole lib tests pass (2387)

+++ crates/litchi-ole/src/xls/border_fill.rs
@@ -222,8 +222,8 @@ pub(crate) fn parse_xf_border_fill(
     let top_color = (border2 & 0x7f) as u16;
     let bottom_color = ((border2 >> 7) & 0x7f) as u16;
     let diagonal_color = ((border2 >> 14) & 0x7f) as u16;
-    let foreground_color = (area & 0x7f) as u16;
-    let background_color = ((area >> 7) & 0x7f) as u16;
+    let foreground_color = area & 0x7f;
+    let background_color = (area >> 7) & 0x7f;
     for (name, color) in [
         ("left border", left_color),
         ("right border", right_color),
